fixed entity re-initialization by switching relationship owners

+++ backend/qservices/src/main/java/ro/pizzeriaq/qservices/data/repository/ProductCategoryRepository.java
@@ -3,5 +3,5 @@ package ro.pizzeriaq.qservices.data.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ro.pizzeriaq.qservices.data.model.ProductCategory;
 
-public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long> {
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Integer> {
 }
